+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changes
 2.21.0 (2025-02-27)
 ^^^^^^^^^^^^^^^^^^^
 * make `AioDeferredRefreshableCredentials` subclass of `DeferredRefreshableCredentials`
+* make `AioSSOCredentialFetcher` subclass of `SSOCredentialFetcher`
 
 2.20.1.dev0 (2025-02-24)
 ^^^^^^^^^^^^^^^^^^^^^^^^